+++ app/src/main/java/com/keylesspalace/tusky/ModalTimelineActivity.kt
@@ -59,7 +59,6 @@ class ModalTimelineActivity : BaseActivity(), ActionButtonActivity, HasSupportFr
                     .replace(R.id.content_frame, TimelineFragment.newInstance(kind, argument))
                     .commit()
         }
-        setupBottomSheet()
     }
 
     override fun getActionButton(): FloatingActionButton? = null
