+++ app/src/main/java/com/keylesspalace/tusky/ViewTagActivity.java
@@ -52,7 +52,6 @@ public class ViewTagActivity extends BaseActivity implements HasSupportFragmentI
             bar.setDisplayHomeAsUpEnabled(true);
             bar.setDisplayShowHomeEnabled(true);
         }
-        setupBottomSheet();
 
         FragmentTransaction fragmentTransaction = getSupportFragmentManager().beginTransaction();
         Fragment fragment = TimelineFragment.newInstance(TimelineFragment.Kind.TAG, hashtag);
