--- app/src/main/java/com/keylesspalace/tusky/interfaces/SearchManager.kt
@@ -1,9 +1,0 @@
-package com.keylesspalace.tusky.interfaces
-
-interface SearchManager {
-    fun onBeginSearch(url: String)
-    fun onEndSearch(url: String)
-    fun getCancelSearchRequested(url: String): Boolean
-    fun getIsSearching(): Boolean
-    fun cancelActiveSearch()
-}
