+++ app/src/main/res/layout/activity_modal_timeline.xml
@@ -24,9 +24,4 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"/>
-    <android.support.design.widget.CoordinatorLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-        <include layout="@layout/item_status_bottom_sheet" />
-    </android.support.design.widget.CoordinatorLayout>
 </android.support.constraint.ConstraintLayout>
