+++ app/src/main/res/layout/activity_view_tag.xml
@@ -16,6 +16,5 @@
         app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
     <include layout="@layout/toolbar_shadow_shim" />
-    <include layout="@layout/item_status_bottom_sheet" />
 
 </android.support.design.widget.CoordinatorLayout>
